+++ cloudauth-intl-20220809/alibabacloud_cloudauth_intl20220809/client.py
@@ -1281,6 +1281,8 @@ class Client(OpenApiClient):
             query['IdFaceQuality'] = request.id_face_quality
         if not UtilClient.is_unset(request.id_ocr_picture_url):
             query['IdOcrPictureUrl'] = request.id_ocr_picture_url
+        if not UtilClient.is_unset(request.id_threshold):
+            query['IdThreshold'] = request.id_threshold
         if not UtilClient.is_unset(request.merchant_biz_id):
             query['MerchantBizId'] = request.merchant_biz_id
         if not UtilClient.is_unset(request.merchant_user_id):
@@ -1334,6 +1336,8 @@ class Client(OpenApiClient):
             query['IdFaceQuality'] = request.id_face_quality
         if not UtilClient.is_unset(request.id_ocr_picture_url):
             query['IdOcrPictureUrl'] = request.id_ocr_picture_url
+        if not UtilClient.is_unset(request.id_threshold):
+            query['IdThreshold'] = request.id_threshold
         if not UtilClient.is_unset(request.merchant_biz_id):
             query['MerchantBizId'] = request.merchant_biz_id
         if not UtilClient.is_unset(request.merchant_user_id):
@@ -1421,6 +1425,8 @@ class Client(OpenApiClient):
             query['FacePictureUrl'] = request.face_picture_url
         if not UtilClient.is_unset(request.id_ocr_picture_url):
             query['IdOcrPictureUrl'] = request.id_ocr_picture_url
+        if not UtilClient.is_unset(request.id_threshold):
+            query['IdThreshold'] = request.id_threshold
         if not UtilClient.is_unset(request.merchant_biz_id):
             query['MerchantBizId'] = request.merchant_biz_id
         if not UtilClient.is_unset(request.merchant_user_id):
@@ -1480,6 +1486,8 @@ class Client(OpenApiClient):
             query['FacePictureUrl'] = request.face_picture_url
         if not UtilClient.is_unset(request.id_ocr_picture_url):
             query['IdOcrPictureUrl'] = request.id_ocr_picture_url
+        if not UtilClient.is_unset(request.id_threshold):
+            query['IdThreshold'] = request.id_threshold
         if not UtilClient.is_unset(request.merchant_biz_id):
             query['MerchantBizId'] = request.merchant_biz_id
         if not UtilClient.is_unset(request.merchant_user_id):
@@ -2023,6 +2031,8 @@ class Client(OpenApiClient):
             query['DocScanMode'] = request.doc_scan_mode
         if not UtilClient.is_unset(request.doc_type):
             query['DocType'] = request.doc_type
+        if not UtilClient.is_unset(request.doc_video):
+            query['DocVideo'] = request.doc_video
         if not UtilClient.is_unset(request.experience_code):
             query['ExperienceCode'] = request.experience_code
         if not UtilClient.is_unset(request.face_picture_url):
@@ -2031,6 +2041,8 @@ class Client(OpenApiClient):
             query['IdFaceQuality'] = request.id_face_quality
         if not UtilClient.is_unset(request.id_spoof):
             query['IdSpoof'] = request.id_spoof
+        if not UtilClient.is_unset(request.id_threshold):
+            query['IdThreshold'] = request.id_threshold
         if not UtilClient.is_unset(request.language_config):
             query['LanguageConfig'] = request.language_config
         if not UtilClient.is_unset(request.merchant_biz_id):
@@ -2039,6 +2051,8 @@ class Client(OpenApiClient):
             query['MerchantUserId'] = request.merchant_user_id
         if not UtilClient.is_unset(request.meta_info):
             query['MetaInfo'] = request.meta_info
+        if not UtilClient.is_unset(request.model):
+            query['Model'] = request.model
         if not UtilClient.is_unset(request.ocr):
             query['Ocr'] = request.ocr
         if not UtilClient.is_unset(request.product_code):
@@ -2102,6 +2116,8 @@ class Client(OpenApiClient):
             query['DocScanMode'] = request.doc_scan_mode
         if not UtilClient.is_unset(request.doc_type):
             query['DocType'] = request.doc_type
+        if not UtilClient.is_unset(request.doc_video):
+            query['DocVideo'] = request.doc_video
         if not UtilClient.is_unset(request.experience_code):
             query['ExperienceCode'] = request.experience_code
         if not UtilClient.is_unset(request.face_picture_url):
@@ -2110,6 +2126,8 @@ class Client(OpenApiClient):
             query['IdFaceQuality'] = request.id_face_quality
         if not UtilClient.is_unset(request.id_spoof):
             query['IdSpoof'] = request.id_spoof
+        if not UtilClient.is_unset(request.id_threshold):
+            query['IdThreshold'] = request.id_threshold
         if not UtilClient.is_unset(request.language_config):
             query['LanguageConfig'] = request.language_config
         if not UtilClient.is_unset(request.merchant_biz_id):
@@ -2118,6 +2136,8 @@ class Client(OpenApiClient):
             query['MerchantUserId'] = request.merchant_user_id
         if not UtilClient.is_unset(request.meta_info):
             query['MetaInfo'] = request.meta_info
+        if not UtilClient.is_unset(request.model):
+            query['Model'] = request.model
         if not UtilClient.is_unset(request.ocr):
             query['Ocr'] = request.ocr
         if not UtilClient.is_unset(request.product_code):